Add support for directly generating bounded text from type level strings
This is nice for places where we want hard-coded bounded text and do not
want to needlessly do error handling.

This drops support for ghc 8.10 and 9.0 as we now use `UnconsSymbol` from
base 4.16. We should discuss if there are any objections to dropping these older
versions and if we think it's worth trying to find a workaround.

Co-Authored-By: Janus <janus@flipstone.com>
Co-Authored-By: Nebula <nebula@flipstone.com>